+++ src/handler.rs
@@ -14,12 +14,10 @@
 //
 #![allow(clippy::large_enum_variant)]
 #![warn(missing_docs)]
-use std::collections::HashMap;
 use std::convert::Infallible;
 use std::error::Error;
 use std::net::{IpAddr, SocketAddr};
 use std::sync::Arc;
-use std::time::Duration;
 
 use ethereum_types::{Address, H256, U256, U64};
 use futures::prelude::*;
@@ -27,31 +25,24 @@ use serde::{Deserialize, Serialize};
 use tokio::sync::mpsc;
 use tokio_stream::wrappers::ReceiverStream;
 use warp::ws::Message;
-use webb::evm::contract::protocol_solidity::fixed_deposit_anchor::Proof;
-use webb::evm::contract::{
-    protocol_solidity::{
-        fixed_deposit_anchor::ExtData, FixedDepositAnchorContract,
-    },
-    tornado::TornadoContract,
-};
 use webb::evm::ethers::{
     contract::ContractError,
     core::k256::SecretKey,
-    middleware::SignerMiddleware,
     providers::Middleware,
     signers::{LocalWallet, Signer},
     types::Bytes,
 };
-use webb::substrate::protocol_substrate_runtime::api::runtime_types::webb_standalone_runtime::Element;
 
 use crate::context::RelayerContext;
 use crate::store::LeafCacheStore;
-use webb::substrate::protocol_substrate_runtime::api::RuntimeApi;
+use crate::tx_relay::evm::anchor::handle_anchor_relay_tx;
+use crate::tx_relay::evm::tornado::handle_tornado_relay_tx;
+use crate::tx_relay::substrate::mixer::handle_substrate_mixer_relay_tx;
 use webb::substrate::subxt::sp_core::Pair;
-use webb::substrate::subxt::DefaultConfig;
-use webb::substrate::subxt::{self, PairSigner, TransactionStatus};
+use webb::substrate::subxt::{self};
+
 /// Type alias for mpsc::Sender<CommandResponse>
-type CommandStream = mpsc::Sender<CommandResponse>;
+pub type CommandStream = mpsc::Sender<CommandResponse>;
 
 /// Sets up a websocket connection.
 ///
@@ -414,449 +405,10 @@ pub async fn handle_evm(
         }
     }
 }
-/// Handler for tornado commands
-///
-/// # Arguments
-///
-/// * `ctx` - RelayContext reference that holds the configuration
-/// * `cmd` - The command to execute
-/// * `stream` - The stream to write the response to
-async fn handle_tornado_relay_tx<'a>(
-    ctx: RelayerContext,
-    cmd: TornadoRelayTransaction,
-    stream: CommandStream,
-) {
-    use CommandResponse::*;
-    let requested_chain = cmd.chain.to_lowercase();
-    let chain = match ctx.config.evm.get(&requested_chain) {
-        Some(v) => v,
-        None => {
-            tracing::event!(
-                target: crate::probe::TARGET,
-                tracing::Level::DEBUG,
-                kind = %crate::probe::Kind::RelayTx,
-                network = ?NetworkStatus::UnsupportedChain,
-            );
-            let _ = stream.send(Network(NetworkStatus::UnsupportedChain)).await;
-            return;
-        }
-    };
-    let supported_contracts: HashMap<_, _> = chain
-        .contracts
-        .iter()
-        .cloned()
-        .filter_map(|c| match c {
-            crate::config::Contract::Tornado(c) => Some(c),
-            _ => None,
-        })
-        .map(|c| (c.common.address, c))
-        .collect();
-    // get the contract configuration
-    let contract_config = match supported_contracts.get(&cmd.contract) {
-        Some(config) => config,
-        None => {
-            tracing::event!(
-                target: crate::probe::TARGET,
-                tracing::Level::DEBUG,
-                kind = %crate::probe::Kind::RelayTx,
-                network = ?NetworkStatus::UnsupportedContract,
-            );
-            let _ = stream
-                .send(Network(NetworkStatus::UnsupportedContract))
-                .await;
-            return;
-        }
-    };
 
-    let wallet = match ctx.evm_wallet(&cmd.chain).await {
-        Ok(v) => v,
-        Err(e) => {
-            tracing::error!("Misconfigured Network: {}", e);
-            tracing::event!(
-                target: crate::probe::TARGET,
-                tracing::Level::DEBUG,
-                kind = %crate::probe::Kind::RelayTx,
-                network = ?NetworkStatus::Misconfigured,
-            );
-            let _ = stream
-                .send(Error(format!("Misconfigured Network: {:?}", cmd.chain)))
-                .await;
-            return;
-        }
-    };
-    // validate the relayer address first before trying
-    // send the transaction.
-    let reward_address = match chain.beneficiary {
-        Some(account) => account,
-        None => wallet.address(),
-    };
-
-    if cmd.relayer != reward_address {
-        tracing::event!(
-            target: crate::probe::TARGET,
-            tracing::Level::DEBUG,
-            kind = %crate::probe::Kind::RelayTx,
-            network = ?NetworkStatus::InvalidRelayerAddress,
-        );
-        let _ = stream
-            .send(Network(NetworkStatus::InvalidRelayerAddress))
-            .await;
-        return;
-    }
-
-    tracing::debug!(
-        "Connecting to chain {:?} .. at {}",
-        cmd.chain,
-        chain.http_endpoint
-    );
-    tracing::event!(
-        target: crate::probe::TARGET,
-        tracing::Level::DEBUG,
-        kind = %crate::probe::Kind::RelayTx,
-        network = ?NetworkStatus::Connecting,
-    );
-    let _ = stream.send(Network(NetworkStatus::Connecting)).await;
-    let provider = match ctx.evm_provider(&cmd.chain).await {
-        Ok(value) => {
-            tracing::event!(
-                target: crate::probe::TARGET,
-                tracing::Level::DEBUG,
-                kind = %crate::probe::Kind::RelayTx,
-                network = ?NetworkStatus::Connected,
-            );
-            let _ = stream.send(Network(NetworkStatus::Connected)).await;
-            value
-        }
-        Err(e) => {
-            let reason = e.to_string();
-            tracing::event!(
-                target: crate::probe::TARGET,
-                tracing::Level::DEBUG,
-                kind = %crate::probe::Kind::RelayTx,
-                network = "Failed",
-                %reason,
-            );
-            let _ =
-                stream.send(Network(NetworkStatus::Failed { reason })).await;
-            tracing::event!(
-                target: crate::probe::TARGET,
-                tracing::Level::DEBUG,
-                kind = %crate::probe::Kind::RelayTx,
-                network = ?NetworkStatus::Disconnected,
-            );
-            let _ = stream.send(Network(NetworkStatus::Disconnected)).await;
-            return;
-        }
-    };
-
-    let client = SignerMiddleware::new(provider, wallet);
-    let client = Arc::new(client);
-    let contract = TornadoContract::new(cmd.contract, client);
-    let denomination = match contract.denomination().call().await {
-        Ok(v) => v,
-        Err(e) => {
-            tracing::event!(
-                target: crate::probe::TARGET,
-                tracing::Level::DEBUG,
-                kind = %crate::probe::Kind::RelayTx,
-                network = "Failed",
-                reason = "Failed to get denomination from contract",
-            );
-            tracing::error!("Misconfigured Contract Denomination: {}", e);
-            let _ = stream
-                .send(Error(format!(
-                    "Misconfigured Contract Denomination: {:?}",
-                    e
-                )))
-                .await;
-            return;
-        }
-    };
-    // check the fee
-    let expected_fee = calculate_fee(
-        contract_config.withdraw_config.withdraw_fee_percentage,
-        denomination,
-    );
-    let (_, unacceptable_fee) = U256::overflowing_sub(cmd.fee, expected_fee);
-    if unacceptable_fee {
-        tracing::error!("Received a fee lower than configuration");
-        let msg = format!(
-            "User sent a fee that is too low {} but expected {}",
-            cmd.fee, expected_fee,
-        );
-        let _ = stream.send(Error(msg)).await;
-        return;
-    }
-
-    let call = contract.withdraw(
-        cmd.proof,
-        cmd.root.to_fixed_bytes(),
-        cmd.nullifier_hash.to_fixed_bytes(),
-        cmd.recipient,
-        cmd.relayer,
-        cmd.fee,
-        cmd.refund,
-    );
-    // Make a dry call, to make sure the transaction will go through successfully
-    // to avoid wasting fees on invalid calls.
-    match call.call().await {
-        Ok(_) => {
-            let _ = stream.send(Withdraw(WithdrawStatus::Valid)).await;
-            tracing::debug!("Proof is valid");
-        }
-        Err(e) => {
-            tracing::error!("Error Client sent an invalid proof: {}", e);
-            let err = into_withdraw_error(e);
-            let _ = stream.send(Withdraw(err)).await;
-            return;
-        }
-    };
-    tracing::trace!("About to send Tx to {:?} Chain", cmd.chain);
-    let tx = match call.send().await {
-        Ok(pending) => {
-            let _ = stream.send(Withdraw(WithdrawStatus::Sent)).await;
-            let tx_hash = *pending;
-            tracing::debug!("Tx is submitted and pending! {}", tx_hash);
-            let result = pending.interval(Duration::from_millis(1000)).await;
-            let _ = stream
-                .send(Withdraw(WithdrawStatus::Submitted { tx_hash }))
-                .await;
-            result
-        }
-        Err(e) => {
-            tracing::error!("Error while sending Tx: {}", e);
-            let err = into_withdraw_error(e);
-            let _ = stream.send(Withdraw(err)).await;
-            return;
-        }
-    };
-    match tx {
-        Ok(Some(receipt)) => {
-            tracing::debug!("Finalized Tx #{}", receipt.transaction_hash);
-            let _ = stream
-                .send(Withdraw(WithdrawStatus::Finalized {
-                    tx_hash: receipt.transaction_hash,
-                }))
-                .await;
-        }
-        Ok(None) => {
-            tracing::warn!("Transaction Dropped from Mempool!!");
-            let _ = stream
-                .send(Withdraw(WithdrawStatus::DroppedFromMemPool))
-                .await;
-        }
-        Err(e) => {
-            let reason = e.to_string();
-            tracing::error!("Transaction Errored: {}", reason);
-            let _ = stream
-                .send(Withdraw(WithdrawStatus::Errored { reason, code: 4 }))
-                .await;
-        }
-    };
-}
-/// Handler for Anchor commands
-///
-/// # Arguments
-///
-/// * `ctx` - RelayContext reference that holds the configuration
-/// * `cmd` - The command to execute
-/// * `stream` - The stream to write the response to
-async fn handle_anchor_relay_tx<'a>(
-    ctx: RelayerContext,
-    cmd: AnchorRelayTransaction,
-    stream: CommandStream,
-) {
-    use CommandResponse::*;
-    let requested_chain = cmd.chain.to_lowercase();
-    let chain = match ctx.config.evm.get(&requested_chain) {
-        Some(v) => v,
-        None => {
-            tracing::warn!("Unsupported Chain: {}", requested_chain);
-            let _ = stream.send(Network(NetworkStatus::UnsupportedChain)).await;
-            return;
-        }
-    };
-    let supported_contracts: HashMap<_, _> = chain
-        .contracts
-        .iter()
-        .cloned()
-        .filter_map(|c| match c {
-            crate::config::Contract::Anchor(c) => Some(c),
-            _ => None,
-        })
-        .map(|c| (c.common.address, c))
-        .collect();
-    // get the contract configuration
-    let contract_config = match supported_contracts.get(&cmd.contract) {
-        Some(config) => config,
-        None => {
-            tracing::warn!("Unsupported Contract: {:?}", cmd.contract);
-            let _ = stream
-                .send(Network(NetworkStatus::UnsupportedContract))
-                .await;
-            return;
-        }
-    };
-
-    let wallet = match ctx.evm_wallet(&cmd.chain).await {
-        Ok(v) => v,
-        Err(e) => {
-            tracing::error!("Misconfigured Network: {}", e);
-            let _ = stream
-                .send(Error(format!("Misconfigured Network: {:?}", cmd.chain)))
-                .await;
-            return;
-        }
-    };
-    // validate the relayer address first before trying
-    // send the transaction.
-    let reward_address = match chain.beneficiary {
-        Some(account) => account,
-        None => wallet.address(),
-    };
-
-    if cmd.relayer != reward_address {
-        let _ = stream
-            .send(Network(NetworkStatus::InvalidRelayerAddress))
-            .await;
-        return;
-    }
-
-    // validate that the roots are multiple of 32s
-    let roots = cmd.roots.to_vec();
-    if roots.len() % 32 != 0 {
-        let _ = stream
-            .send(Withdraw(WithdrawStatus::InvalidMerkleRoots))
-            .await;
-        return;
-    }
-
-    tracing::debug!(
-        "Connecting to chain {:?} .. at {}",
-        cmd.chain,
-        chain.http_endpoint
-    );
-    let _ = stream.send(Network(NetworkStatus::Connecting)).await;
-    let provider = match ctx.evm_provider(&cmd.chain).await {
-        Ok(value) => {
-            let _ = stream.send(Network(NetworkStatus::Connected)).await;
-            value
-        }
-        Err(e) => {
-            let reason = e.to_string();
-            let _ =
-                stream.send(Network(NetworkStatus::Failed { reason })).await;
-            let _ = stream.send(Network(NetworkStatus::Disconnected)).await;
-            return;
-        }
-    };
-
-    let client = SignerMiddleware::new(provider, wallet);
-    let client = Arc::new(client);
-    let contract = FixedDepositAnchorContract::new(cmd.contract, client);
-    let denomination = match contract.denomination().call().await {
-        Ok(v) => v,
-        Err(e) => {
-            tracing::error!("Misconfigured Contract Denomination: {}", e);
-            let _ = stream
-                .send(Error(format!(
-                    "Misconfigured Contract: {:?}",
-                    cmd.contract
-                )))
-                .await;
-            return;
-        }
-    };
-    // check the fee
-    let expected_fee = calculate_fee(
-        contract_config.withdraw_config.withdraw_fee_percentage,
-        denomination,
-    );
-    let (_, unacceptable_fee) = U256::overflowing_sub(cmd.fee, expected_fee);
-    if unacceptable_fee {
-        tracing::error!("Received a fee lower than configuration");
-        let msg = format!(
-            "User sent a fee that is too low {} but expected {}",
-            cmd.fee, expected_fee,
-        );
-        let _ = stream.send(Error(msg)).await;
-        return;
-    }
-
-    let ext_data = ExtData {
-        refresh_commitment: cmd.refresh_commitment.to_fixed_bytes(),
-        recipient: cmd.recipient,
-        relayer: cmd.relayer,
-        fee: cmd.fee,
-        refund: cmd.refund,
-    };
-    let proof = Proof {
-        roots: roots.into(),
-        proof: cmd.proof,
-        nullifier_hash: cmd.nullifier_hash.to_fixed_bytes(),
-        ext_data_hash: cmd.ext_data_hash.to_fixed_bytes(),
-    };
-    tracing::trace!(?proof, ?ext_data, "Client Proof");
-    let call = contract.withdraw(proof, ext_data);
-    // Make a dry call, to make sure the transaction will go through successfully
-    // to avoid wasting fees on invalid calls.
-    match call.call().await {
-        Ok(_) => {
-            let _ = stream.send(Withdraw(WithdrawStatus::Valid)).await;
-            tracing::debug!("Proof is valid");
-        }
-        Err(e) => {
-            tracing::error!("Error Client sent an invalid proof: {}", e);
-            let err = into_withdraw_error(e);
-            let _ = stream.send(Withdraw(err)).await;
-            return;
-        }
-    };
-    tracing::trace!("About to send Tx to {:?} Chain", cmd.chain);
-    let tx = match call.send().await {
-        Ok(pending) => {
-            let _ = stream.send(Withdraw(WithdrawStatus::Sent)).await;
-            let tx_hash = *pending;
-            tracing::debug!(%tx_hash, "Tx is submitted and pending!");
-            let result = pending.interval(Duration::from_millis(1000)).await;
-            let _ = stream
-                .send(Withdraw(WithdrawStatus::Submitted { tx_hash }))
-                .await;
-            result
-        }
-        Err(e) => {
-            tracing::error!("Error while sending Tx: {}", e);
-            let err = into_withdraw_error(e);
-            let _ = stream.send(Withdraw(err)).await;
-            return;
-        }
-    };
-    match tx {
-        Ok(Some(receipt)) => {
-            tracing::debug!("Finalized Tx #{}", receipt.transaction_hash);
-            let _ = stream
-                .send(Withdraw(WithdrawStatus::Finalized {
-                    tx_hash: receipt.transaction_hash,
-                }))
-                .await;
-        }
-        Ok(None) => {
-            tracing::warn!("Transaction Dropped from Mempool!!");
-            let _ = stream
-                .send(Withdraw(WithdrawStatus::DroppedFromMemPool))
-                .await;
-        }
-        Err(e) => {
-            let reason = e.to_string();
-            tracing::error!("Transaction Errored: {}", reason);
-            let _ = stream
-                .send(Withdraw(WithdrawStatus::Errored { reason, code: 4 }))
-                .await;
-        }
-    };
-}
-
-fn into_withdraw_error<M: Middleware>(e: ContractError<M>) -> WithdrawStatus {
+pub fn into_withdraw_error<M: Middleware>(
+    e: ContractError<M>,
+) -> WithdrawStatus {
     // a poor man error parser
     // WARNING: **don't try this at home**.
     let msg = format!("{}", e);
@@ -905,149 +457,9 @@ pub async fn handle_substrate<'a>(
         }
     }
 }
-/// Handler for Substrate Mixer commands
-///
-/// # Arguments
-///
-/// * `ctx` - RelayContext reference that holds the configuration
-/// * `cmd` - The command to execute
-/// * `stream` - The stream to write the response to
-async fn handle_substrate_mixer_relay_tx<'a>(
-    ctx: RelayerContext,
-    cmd: MixerRelayTransaction,
-    stream: CommandStream,
-) {
-    use CommandResponse::*;
 
-    let root_element = Element(cmd.root);
-    let nullifier_hash_element = Element(cmd.nullifier_hash);
-
-    let requested_chain = cmd.chain.to_lowercase();
-    let maybe_client = ctx
-        .substrate_provider::<DefaultConfig>(&requested_chain)
-        .await;
-    let client = match maybe_client {
-        Ok(c) => c,
-        Err(e) => {
-            tracing::error!("Error while getting Substrate client: {}", e);
-            let _ = stream.send(Error(format!("{}", e))).await;
-            return;
-        }
-    };
-    let api = client.to_runtime_api::<RuntimeApi<DefaultConfig, subxt::DefaultExtra<DefaultConfig>>>();
-
-    let pair = match ctx.substrate_wallet(&cmd.chain).await {
-        Ok(v) => v,
-        Err(e) => {
-            tracing::error!("Misconfigured Network: {}", e);
-            let _ = stream
-                .send(Error(format!("Misconfigured Network: {:?}", cmd.chain)))
-                .await;
-            return;
-        }
-    };
-
-    let signer = PairSigner::new(pair);
-
-    let withdraw_tx = api
-        .tx()
-        .mixer_bn254()
-        .withdraw(
-            cmd.id,
-            cmd.proof,
-            root_element,
-            nullifier_hash_element,
-            cmd.recipient,
-            cmd.relayer,
-            cmd.fee,
-            cmd.refund,
-        )
-        .sign_and_submit_then_watch(&signer)
-        .await;
-    let mut event_stream = match withdraw_tx {
-        Ok(s) => s,
-        Err(e) => {
-            tracing::error!("Error while sending Tx: {}", e);
-            let _ = stream.send(Error(format!("{}", e))).await;
-            return;
-        }
-    };
-
-    // Listen to the withdraw transaction, and send information back to the client
-    loop {
-        let maybe_event = event_stream.next().await;
-        let event = match maybe_event {
-            Some(Ok(v)) => v,
-            Some(Err(e)) => {
-                tracing::error!("Error while watching Tx: {}", e);
-                let _ = stream.send(Error(format!("{}", e))).await;
-                return;
-            }
-            None => break,
-        };
-        match event {
-            TransactionStatus::Broadcast(_) => {
-                let _ = stream.send(Withdraw(WithdrawStatus::Sent)).await;
-            }
-            TransactionStatus::InBlock(info) => {
-                tracing::debug!(
-                    "Transaction {:?} made it into block {:?}",
-                    info.extrinsic_hash(),
-                    info.block_hash()
-                );
-                let _ = stream
-                    .send(Withdraw(WithdrawStatus::Submitted {
-                        tx_hash: H256::from_slice(
-                            info.extrinsic_hash().as_bytes(),
-                        ),
-                    }))
-                    .await;
-            }
-            TransactionStatus::Finalized(info) => {
-                tracing::debug!(
-                    "Transaction {:?} finalized in block {:?}",
-                    info.extrinsic_hash(),
-                    info.block_hash()
-                );
-                let _has_event = match info.wait_for_success().await {
-                    Ok(_) => {
-                        // TODO: check if the event is actually a withdraw event
-                        true
-                    }
-                    Err(e) => {
-                        tracing::error!("Error while watching Tx: {}", e);
-                        let _ = stream.send(Error(format!("{}", e))).await;
-                        false
-                    }
-                };
-                let _ = stream
-                    .send(Withdraw(WithdrawStatus::Finalized {
-                        tx_hash: H256::from_slice(
-                            info.extrinsic_hash().as_bytes(),
-                        ),
-                    }))
-                    .await;
-            }
-            TransactionStatus::Dropped => {
-                tracing::warn!("Transaction dropped from the pool");
-                let _ = stream
-                    .send(Withdraw(WithdrawStatus::DroppedFromMemPool))
-                    .await;
-            }
-            TransactionStatus::Invalid => {
-                let _ = stream
-                    .send(Withdraw(WithdrawStatus::Errored {
-                        reason: "Invalid".to_string(),
-                        code: 4,
-                    }))
-                    .await;
-            }
-            _ => continue,
-        }
-    }
-}
 /// Calculates the fee for a given transaction
-fn calculate_fee(fee_percent: f64, principle: U256) -> U256 {
+pub fn calculate_fee(fee_percent: f64, principle: U256) -> U256 {
     let mill_fee = (fee_percent * 1_000_000.0) as u32;
     let mill_u256: U256 = principle * (mill_fee);
     let fee_u256: U256 = mill_u256 / (1_000_000);
